+++ src/Http/Middleware/Authenticate.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace PavelMironchik\LaravelBackupPanel\Http\Middleware;
+namespace Trungdev05\LaravelBackupPanel\Http\Middleware;
 
 use Closure;
 use Illuminate\Http\Request;
-use PavelMironchik\LaravelBackupPanel\LaravelBackupPanel;
 use Symfony\Component\HttpFoundation\Response;
+use Trungdev05\LaravelBackupPanel\LaravelBackupPanel;
 
 class Authenticate
 {