+++ src/Http/Requests/ShowBackupPanelRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace PavelMironchik\LaravelBackupPanel\Http\Requests;
+namespace Trungdev05\LaravelBackupPanel\Http\Requests;
 
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
-use PavelMironchik\LaravelBackupPanel\Rules\BackupDisk;
+use Trungdev05\LaravelBackupPanel\Rules\BackupDisk;
 
 final class ShowBackupPanelRequest extends FormRequest
 {